Motivation for this patch: UCSC track hubs require bigBed files. segway 1.2 cannot generate bigBed files (see issue #22).
Patched layer.py to make sure that the segment added as a pad has length 1 instead of 0.

Updated segway.layered.bed and segway.layered.bb in test/segway_layer/touchstone. Tests pass.

Visual inspection of the old and new outputs in the browser shows that this creates a segment of length 1 at the first chromosomal coordinate. But I don't see any way around this, as bedToBigBed requires bedSizes > 0.

Another side effect is that I hard coded the `-b` option in the test, which means that 1 of the 2 tests will fail on machines where the `bedToBigBed` executable is not installed.