Commit initial implementation and tests of GMTK error reports
Adds a simplebadinput test to check for proper termination on a GMTK job
submitted error. Adds documentation specifying how to better debug GMTK
errors. Adds support to catch out of memory errors in GTMK versions 1.4.3 or
above. Adds a better error message for errors that are not due to being out of
memory.